// work/AzureDevOps/angular/app/src/app/modules/produtos/sub-modules/e-contrato/core/models/contratos/veiculo.model.ts
export class Veiculo {
    id?: number;
    chassi: string;
    placa: string;
    ufPlaca: string;
    anoFabricacao: number;
    anoModelo: number;
    renavam: string;
    numeroRestricao: number;
    marca: string;
    modelo: string;
    emplacado: boolean;
    remarcado: boolean;
    especie: number;
    cor: string;
    podeEditar: boolean;
}